Add guard clause for InferenceError in slice inference
Fixes #5. It's unclear if InferenceError is supposed to bubble up to the
surface as it seems to be an implementation detail of astroid which
should be caught and trigger the return of Uninferable. For now we'll
catch it and pretend it was Uninferable all along and clarify the
expected behaviour with astroid in the future.