Fix KeyError when attempting to alias missing module
Ref #6. Maybe we should revisit the behavior and invariants around
aliasing imports. This fixes the previous assumption that `import_`
would always succeed: even though the assert only checks that new_scope
is a superset of the current scope, the `alias is not None` branch
assumes that `new_scope` is a strict superset and so can raise
`KeyError` when it's not (like in the case of a failed import just
addressed).