Update tests to warn on missing imports
If the linter is unable to import what it expects to be a protobuf
module definition, it defers to conservative behaviour and won't warn
(obviously it is unable to check any definitions). This change in
behaviour has caused some differences between testing environments in
the past, so we explicitly fail on this behaviour when testing, since
we should be able to import every module as we control the testing
environment.