Replace field tracking paradigm with AST transform and inference
Large amounts of complexity in the previous approach (and attempts to
re-write it in .evaluation) surrounded thet tracking of scopes,
renaming, etc. This is better served by the extensive work done with
astroid inference. Here we leverage an AST transform to replace message
types generated by reflection with a plain class definition, enabling
astroid to infer these values and adding an annotation to enable our
checker to later recognise these substitutions.

I am unhappy with the current use of exec() however this approach is
much more promising than previous attempts.