Refactor ClassDef annotation in .transform
The whole effort of this change boils down to it being a struggle to
associate the protobuf descriptors to the generated classes. Astroid
wants to parse things from the top down (building the whole tree) so
that the relationship between parent and child nodes can be recorded but
we lose the relationship between message descriptors and nested types by
waiting until the whole class string is built.