Fix inference of non-nested top-level message types
When generating the ClassDef nodes, they are created in an empty module
by astroid.parse. Adding these to the Module.locals meant that that the
appropriate ClassDef nodes would be retrieved when inferring an
ImportFrom statement, but when inferring the value of a member on an
instantiated class astroid would walk up to the Node.root() (getting the
bare module from its creation, not the _pb2 module). Fixing this link
fixed the inference of name resolution for other members in the module.