Update ambiguous inference to match pylint behaviour
Ref #44. When inferring the type of messages, astroid doesn't attempt to
resolve ambiguous cases, instead preferring to return a set of possible
values. For example:

    node = astroid.extract_node("""
        items = [1, 2, 3]
        val = 1
        if 1 + 1 == 2:
            val = 'blue'
        val
    """)
    print(list(node.infer()))
    # [<Const.int l.3 at 0x...>, <Const.str l.5 at 0x...>]

For built-in checkers like pylint.checkers.typecheck they resolve this
by checking if any object, say, has the missing member and only raising
the warning if *none* of them do. Previously, we only checked the first
inference result that looked like a protobuf class. This change updates
the behaviour to match the pylint behaviour.