Fix global state with protobuf DescriptorPool
This singleton is what was causing problems with "redefinition" of
message descriptors again and again throughout our test cases, and let
to the "package test_name" shenanigans. This worked well enough but is
no good if we actually wanted to test behaviours in the case where we
are explicitly defining messages without a package (and those names
clash).