Fix no-member by marking transformed class with __getattr__
Since checker priority was removed, it no longer makes sense to mark the
relevant lines with "disable: no-member", and instead just give up and
mark the generated class as dynamic so pylint/typecheck:_emit_no_member
ignores the result. As of Pylint 3.x, the checker execution order has
changed so ProtobufDescriptorChecker was executing after the builtin
typecheck checker, and it was too late to prevent no-member being
emitted.

TODO: in a future change, remove the `_disable` method.